Revert "Add reconnecting ptys (#23)"
This partially reverts commit 91201718149964a06c6350679d8cda0161edf887.

The new method using screen will not share processes which is a
fundamental shift so I think it will be easier to start from scratch.

Even though we could keep the UUID check I removed it because it seems
cool that you could create your own sessions in the terminal then
connect to them in the browser (or vice-versa).