Implement reconnections with screen (#25)
* Refactor reconnect test to support sub-tests

Going to add an alternate screen test next.

* Revert "Add reconnecting ptys (#23)"

This partially reverts commit 91201718149964a06c6350679d8cda0161edf887.

The new method using screen will not share processes which is a
fundamental shift so I think it will be easier to start from scratch.

Even though we could keep the UUID check I removed it because it seems
cool that you could create your own sessions in the terminal then
connect to them in the browser (or vice-versa).

* Add test for alternate screen

The output test waits for EOF; modify that behavior so we can check that
certain strings are not displayed *without* waiting for the timeout.
This means to be accurate we should always check for output that should
exist after the output that should not exist would have shown up.

* Add timeout flag to dev client

This makes it easier to test reconnects manually.

* Add size to initial connection

This way you do not need a subsequent resize and we can have the right
size from the get-go.

* Prevent prompt from rendering twice in tests

* Add Nix flake

* Propagate process close error

* Implement reconnecting TTY with screen

* Encapsulate session logic

* Localize session map

* Consolidate test scaffolding into helpers

I think this helps make the tests a bit more concise.

* Test many connections at once

* Fix errors not propagating through web socket close

Since the server closed the socket the caller has no chance to close
with the right code and reason.

Also abnormal closure is not a valid close code.

* Fix test flake in reading output

Without waiting for the copy you can sometimes get "file already
closed".  I guess process.Wait must have some side effect.