Add a close reason
Trying to figure out why a session is prematurely closing.

To do this I am setting the error via setState (so I can add the reason
at the close call site) rather than checking for a nil error and
returning it in the Attach.  Alternatively I was thinking of adding a
reason arg to setState but only the close state needs a reason and
ultimately it was transformed into the error anyway so might as well do
it earlier and skip a step.