Fix attach resize (#29)
* StatusAbnormalClosure cannot be set

So says the warning emitted by the websocket library every time we try
to set StatusAbnormalClosure.

* Refactor screen sessions

Previously we would spawn the daemon via -Dm then attach with -x once it
was ready.  This had a flaw: the daemon starts with a hardcoded 24x80
size and when the attach comes in it resizes and leaves a bunch of
confusing whitespace above your prompt.

Now we skip the daemon spawn and go straight for the attach but with the
addition of -RR which lets screen spawn the daemon for us if it does
not yet exist.

Consequences:

1. We can only allow one attach at a time because screen has no problem
   creating multiple daemons with the same name.
2. IDs cannot overlap since screen will do partial matching when we do
   not include the PID which we no longer have.
3. We do not know when the daemon exits so cleanup only happens on the
   timeout.
4. We have to kill the session by sending the quit command through
   screen.  When we do this it is possible the session is already dead.
5. If the daemon exits and the user reconnects before the timeout the
   daemon will be respawned all while the Go program remains blissfully
   unaware assuming it has been up this whole time.  Does not change
   anything in practice, just a bit different in terms of underlying
   architecture.

In some ways this new architecture is actually simpler with roughly the
same functionality but it does not support concurrent attaches and has a
greater danger of desyncing from screen's own state.

* Log expected/unexpected and strip ansi from tests

Hopefully makes it a bit easier to see what is going on.

* Add a close reason

Trying to figure out why a session is prematurely closing.

To do this I am setting the error via setState (so I can add the reason
at the close call site) rather than checking for a nil error and
returning it in the Attach.  Alternatively I was thinking of adding a
reason arg to setState but only the close state needs a reason and
ultimately it was transformed into the error anyway so might as well do
it earlier and skip a step.

* Increase test timeout

This was canceling while the reconnect test was running.

* Decrease number of concurrent test sessions

Having too many seems to be causing some to exit unexpectedly.